refactor: Reuse sendMessage function
